reorganise tests and add to the readme
