+++ README.md
@@ -90,6 +90,26 @@ Use the planning dir for plans
    The app will be available at [http://localhost:8080](http://localhost:8080) {or [http://localhost:3000](http://localhost:3000) with browser-sync}.
 
 
+## Running Tests
+
+Tests live in the [tests/](tests/) directory and use Python's built-in `unittest`. From the project root:
+
+- Run everything:
+
+  ```bash
+  python -m unittest
+  ```
+
+- Run just the unit tests or just the Selenium system tests:
+
+  ```bash
+  python -m unittest tests.unittests.test_app
+  python -m unittest tests.systemtests.test_public
+  ```
+
+The system tests use Selenium with Firefox (geckodriver required) and run headless by default. Set `HEADLESS=0` in your .env to watch the browser, e.g. `HEADLESS=0 python -m unittest`.
+
+
 ## Docs
 
 Go to the /docs directory for explantions for how the app works.
