+++ tests/unittests/test_friends.py
